+++ plugins/eslint-plugin-omnicalculator/rules/nullish-utils.js
